Refactor control module definitions to make code more readable
